+++ apps/api/src/app/activities/activities.service.ts
@@ -193,7 +193,8 @@ export class ActivitiesService {
       userId: data.userId
     });
 
-    let account: Prisma.AccountCreateNestedOneWithoutActivitiesInput;
+    let account:
+      Prisma.AccountCreateNestedOneWithoutActivitiesInput | undefined;
 
     if (data.accountId) {
       account = {
@@ -1031,8 +1032,8 @@ export class ActivitiesService {
               },
               {
                 OR: [
-                  { assetProfileOverrides: { is: null } },
-                  { assetProfileOverrides: { assetClass: null } }
+                  { assetProfileOverrides: { assetClass: null } },
+                  { assetProfileOverrides: { is: null } }
                 ]
               }
             ]