+++ apps/api/src/services/data-provider/financial-modeling-prep/financial-modeling-prep.service.ts
@@ -59,7 +59,8 @@ export class FinancialModelingPrepService
     Macau: 'MO',
     'Russian Federation': 'RU',
     'Taiwan (Province of China)': 'TW',
-    Turkey: 'TR'
+    Turkey: 'TR',
+    'Virgin Islands (British)': 'VG'
   };
 
   private readonly logger = new Logger(FinancialModelingPrepService.name);
@@ -726,11 +727,11 @@ export class FinancialModelingPrepService
   }
 
   private parseAssetClass(profile: any): {
-    assetClass: AssetClass;
-    assetSubClass: AssetSubClass;
+    assetClass: AssetClass | undefined;
+    assetSubClass: AssetSubClass | undefined;
   } {
-    let assetClass: AssetClass;
-    let assetSubClass: AssetSubClass;
+    let assetClass: AssetClass | undefined;
+    let assetSubClass: AssetSubClass | undefined;
 
     if (profile) {
       if (profile.isEtf) {