+++ libs/common/src/lib/interfaces/responses/public-portfolio-response.interface.ts
@@ -17,14 +17,16 @@ export interface PublicPortfolioResponse extends PublicPortfolioResponseV1 {
     | 'dateOfFirstActivity'
     | 'markets'
     | 'netPerformancePercentWithCurrencyEffect'
-    | 'valueInBaseCurrency'
     | 'valueInPercentage'
   >[];
   latestActivities: (Pick<
     Order,
     'currency' | 'date' | 'fee' | 'quantity' | 'type' | 'unitPrice'
   > & {
-    assetProfile?: EnhancedAssetProfile;
+    assetProfile: Pick<
+      EnhancedAssetProfile,
+      'currency' | 'dataSource' | 'name' | 'symbol'
+    >;
     value: number;
     valueInBaseCurrency: number;
   })[];